refactor: rename MailOrPhoneForm to EmailForm
- Rename MailOrPhoneForm component to EmailForm
- Rename mail-or-phone-form.tsx to email-form.tsx
- Update component props type from MailOrPhoneFormProps to EmailFormProps
- Update import in app/page.tsx
